+++ core/scripts/full_build_and_run.sh
@@ -1,6 +1,5 @@
 set -euo pipefail
 script_dirpath="$(cd "$(dirname "${0}")" && pwd)"
 
-bash "${script_dirpath}/build_initializer_binary.sh"
-bash "${script_dirpath}/build_api_image.sh"
+bash "${script_dirpath}/build_images.sh"
 bash "${script_dirpath}/run.sh" ${*:-}